Refactor Character class to change armor type from float to int in constructors and member variables

+++ src/Character.cpp
@@ -4,11 +4,11 @@
 
 #include "Character.h"
 
-Character::Character() : name(""), health(0), maxHp(0), armor(0.0f)
+Character::Character() : name(""), health(0), maxHp(0), armor(0)
 {
 }
 
-Character::Character(const std::string &name, int health, float armor) : name(name), health(health), maxHp(health), armor(armor)
+Character::Character(const std::string &name, int health, int armor) : name(name), health(health), maxHp(health), armor(armor)
 {
     if (maxHp < 0) maxHp = 0;
     if (health < 0) health = 0;
